CCD-2378: Re-enable Fortify scan (#143)
* CCD-2378: Re-enable Fortify scan

- Changed fortify release id in fortify-client.properties to new value provided by fortify support.  This should resolve the "The entitlementPreferenceType is not available" error in the Fortify Scan stage of the nightly build pipeline.
- Uncommented Fortify scan steps in Jenkinsfile_nightly to re-enable Fortify scan

* CCD-2378: Temporarily ignore Fortify scan results

Temporarily changed fortifyScan task in build.gradle to ignore results of scan so build doesn't fail.  This will be reverted once Fortify issues identified as false positives have been accepted as such.

* CCD-2378: Undo change to ignore Fortify scan results

Changed fortifyScan task in build.gradle so that it no longer ignores the results of the scan.  Rather than removing the change it has been commented out so that it can be re-instated in the future should the need arise.